main.html: Create icon view for application, emulate style
Instead of JSON, return view of applications as tiled icons with
truncated titles. Whole container (icon and title) should link to the
app's page in the Apple store.

Eventually, these links should be replaced with endpoint calls to
actually install the app, or, at the very least, open it in the App
Store program.

Additionally, store descriptions in hidden divs, to become part of
design later.